Commit env vars to a tracked rs file, fix Cargo.toml issues, fix array size length
Some vars in our use case will not change, and to simplify deployment we
will build them straight into the source of the application. Cargo.toml
had issues with version being 2023 (which is an invalid Cargo version
specifier), and needed to be changed back to the current latest (2021).
Finally, wordlist.rs size specifier N needed to be changed to 201, which
is the exact amount of words in the array.